requirements.txt
numpy == 1.21.5
torch==2.4.1
torchvision==0.19.1
git+https://github.com/openai/CLIP.git
pytorch-lightning==1.8.1
timm==1.0.9
matplotlib==3.8.4  "numpy == 1.21.5"
h5py==3.7.0
open3d==0.18.0     "numpy == 1.21.5"
networkx==2.8.8
pyvisgraph==0.2.1
gdown==5.2.0
hydra-core==1.3.2
#shapely==1.8.0
